+++ sources/led-drivers/spi/DriverSpiLpd8806.cpp
@@ -100,7 +100,7 @@ int DriverSpiLpd8806::writeFiniteColors(const std::vector<ColorRgb>& ledValues)
 	if (_ledCount != ledValues.size())
 	{
 		Warning(_log, "Lpd8806 led's number has changed (old: %d, new: %d). Rebuilding buffer.", _ledCount, ledValues.size());
-		_ledCount = ledValues.size();
+		_ledCount = static_cast<uint>(ledValues.size());
 
 		_ledBuffer.resize(0, 0x00);
 
@@ -121,7 +121,7 @@ int DriverSpiLpd8806::writeFiniteColors(const std::vector<ColorRgb>& ledValues)
 	}
 
 	// Write the data
-	return writeBytes(_ledBuffer.size(), _ledBuffer.data());
+	return writeBytes(static_cast<unsigned int>(_ledBuffer.size()), _ledBuffer.data());
 }
 
 LedDevice* DriverSpiLpd8806::construct(const QJsonObject& deviceConfig)
